Performance enhances (#88)
* Changed bulk write to update many on block table sync complete

* Iterate through the transaction list once on block puller,
Changed block indexes to descending to better fit most of the lookup requests.

* Added const for hard coded address string

* Changed puller to perform multiple fetches from client on IBD

* Added limit on blocking collection in case the pull is much faster than the insert

* Converted block index to Uint on the mongo db

* Throw if the blocking collection task had an exception

* Fix CRS custom config

* Added configurations for puller, consolidated logic for tasks in puller (IBD) Added abort for cancellation token in puller after blocking collection

* Some minor improvments

* rename blocking collection

* Fixed a bug in the number of tasks and size of the list of tasks

* Moved the call to get the tip out of the tasks and into the runner so it's only called once and exit if we are at the tip

* Changed to fit the value of int from mongo to avoid overflow

* Added guard for empty results in the rewind lookup

* Darin the blocking colleciion on reorg

+++ src/Blockcore.Indexer.Core/Storage/Mongo/Types/AddressHistoryComputedTable.cs
@@ -15,6 +15,6 @@ namespace Blockcore.Indexer.Core.Storage.Mongo.Types
 
       public long Position { get; set; }
 
-      public long BlockIndex { get; set; }
+      public uint BlockIndex { get; set; }
    }
 }
