+++ src/Blockcore.Indexer.Core/Storage/Mongo/Types/TransactionBlockTable.cs
@@ -2,7 +2,7 @@ namespace Blockcore.Indexer.Core.Storage.Mongo.Types
 {
    public class TransactionBlockTable
    {
-      public long BlockIndex { get; set; }
+      public uint BlockIndex { get; set; }
 
       public string TransactionId { get; set; }
    }
